fix (BookmarkDao) : deleteBookmark 쿼리 수정 및 pageIndex type 을 Int 로 변경

+++ app/src/main/java/com/ejooyoung/pdf_reader/base/repository/BookmarkRepository.kt
@@ -6,9 +6,9 @@ import io.reactivex.rxjava3.core.Flowable
 
 interface BookmarkRepository {
     fun selectAllBookmarks(bookGuid: String): Flowable<List<Bookmark>>
-    fun insertBookmark(pageIdx: Long, bookGuid: String): Completable
-    fun deleteBookmark(pageIdx: Long, bookGuid: String): Completable
+    fun insertBookmark(pageIdx: Int, bookGuid: String): Completable
+    fun deleteBookmark(pageIdx: Int, bookGuid: String): Completable
     fun deleteBookmark(bookGuid: String): Completable
     fun updateBookmark(bookmark: Bookmark): Completable
-    fun isBookmarkedPage(bookGuid: String, pageIdx: Long): Flowable<Boolean>
+    fun isBookmarkedPage(bookGuid: String, pageIdx: Int): Flowable<Boolean>
 }
